refactor(geocoding): refactor imports and change mapzen config file to config dir
