+++ src/app/services/geocode.services/mapzen-geocode.service.spec.ts
@@ -3,11 +3,13 @@
 import {TestBed, inject, fakeAsync} from '@angular/core/testing';
 import {HttpModule, Http, BaseRequestOptions, ResponseOptions, Response} from '@angular/http';
 import {MockBackend} from '@angular/http/testing';
+
 import 'rxjs/add/observable/of';
 import 'rxjs/add/operator/map';
 import 'rxjs/add/operator/catch';
+
 import {MapzenGeocodeService} from './mapzen-geocode.service';
-import {MAPZEN_BASE_URL} from './mapzen.config';
+import {MAPZEN_BASE_URL} from '../../../config/mapzen.config';
 
 describe('MapzenGeocodeService', () => {
   beforeEach(() => {
