--- src/app/services/geocode.services/mapzen.config.ts
@@ -1,9 +1,0 @@
-export const MAPZEN_API_KEY = 'mapzen-YJqMyuT';
-export const MAPZEN_BASE_URL = 'https://search.mapzen.com/v1/search';
-export const SEARCH_PARAMS = {
-  'boundary.country': 'ES',
-  'sources': 'oa',
-  'layers': 'address',
-  'size': '1',
-  'api_key': MAPZEN_API_KEY
-};
